pyproject.toml
[project]
name = "streaming-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.3",
    "ollama>=0.6.1",
    "openai>=2.16.0",
    "transformers==4.51.3",
    "opencv-python>=4.9.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "pillow>=10.2.0",
    "numpy>=1.24.0",
    "lap>=0.5.12",
    "dotenv>=0.9.9",
    "ultralytics>=8.4.12",
    "timm>=1.0.24",
    "einops>=0.8.2",
    "accelerate>=1.12.0",
    "uvicorn>=0.40.0",
]
